fix test, more refactoring
